Remove dependency to MainActivity in ProfileFragment
